requirements.txt
beautifulsoup4==4.12.3
httpx==0.27.2
lxml==5.3.0
selectolax==0.3.21
tqdm==4.66.5
tenacity==9.0.0
pydantic==2.9.2
playwright==1.47.0
python-dotenv==1.0.1
PyYAML==6.0.2
